fix u32-intrinsics

+++ lib/c/lib_intrinsics.h
@@ -33,7 +33,7 @@
 #else
 
 #define Lib_IntTypes_Intrinsics_add_carry_u32(x1, x2, x3, x4) \
-  (_addcarry_u32(x1, x2, x3, (long unsigned int *) x4))
+  (_addcarry_u32(x1, x2, x3, (unsigned int *) x4))
 
 #define Lib_IntTypes_Intrinsics_add_carry_u64(x1, x2, x3, x4) \
   (_addcarry_u64(x1, x2, x3, (long long unsigned int *) x4))
@@ -49,7 +49,7 @@
   (__GNUC__ < 7 || (__GNUC__ == 7 && (__GNUC_MINOR__ < 2)))
 
 #define Lib_IntTypes_Intrinsics_sub_borrow_u32(x1, x2, x3, x4) \
-  (_subborrow_u32(x1, x3, x2, (long long unsigned int *) x4))
+  (_subborrow_u32(x1, x3, x2, (unsigned int *) x4))
 
 #define Lib_IntTypes_Intrinsics_sub_borrow_u64(x1, x2, x3, x4) \
   (_subborrow_u64(x1, x3, x2, (long long unsigned int *) x4))
@@ -57,7 +57,7 @@
 #else
 
 #define Lib_IntTypes_Intrinsics_sub_borrow_u32(x1, x2, x3, x4)  \
-  (_subborrow_u32(x1, x2, x3, (long long unsigned int *) x4))
+  (_subborrow_u32(x1, x2, x3, (unsigned int *) x4))
 
 #define Lib_IntTypes_Intrinsics_sub_borrow_u64(x1, x2, x3, x4)  \
   (_subborrow_u64(x1, x2, x3, (long long unsigned int *) x4))
